scrapers/requirements.txt
# Scraper dependencies (used by daily-meetings-sync GitHub Action)
# Kept separate because Playwright is heavy (~150MB) and only needed here
playwright>=1.40.0
supabase>=2.5.0
python-dotenv>=1.0.0
anthropic>=0.18.0
pdfplumber>=0.10.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
